Add crud

+++ src/main/java/com/talentreef/interviewquestions/takehome/respositories/WidgetRepository.java
@@ -13,6 +13,16 @@ public class WidgetRepository {
 
   private List<Widget> table = new ArrayList<>();
 
+  public WidgetRepository(List<Widget> table) {
+    this.table = table;
+    setWidgetTestData(table);
+  }
+  private void setWidgetTestData(List<Widget> table){
+    table.add(Widget.builder().name("Widgette Nielson").description("Keeps a diary").price(10.2).build());
+    table.add(Widget.builder().name("Naomi Ceron").description("Takes home a challenge").price(99.99).build());
+    table.add(Widget.builder().name("Jorge Reyes").description("Is a great dev").price(10).build());
+  }
+
   public List<Widget> deleteById(String name) {
     this.table = table.stream()
                       .filter((Widget widget) -> !name.equals(widget.getName()))
@@ -32,7 +42,9 @@ public class WidgetRepository {
   }
 
   public Widget save(Widget widget) {
-    deleteById(widget.getName());
+    if(findById(widget.getName()).isPresent()){
+      deleteById(widget.getName());
+    }
     table.add(widget);
     return widget;
   }
